Remove previous changed for maximum_heap_size for upgrade; fix logstash
The previous changes applied in commit [1] are removed. The sfconfig
on upgrade in CI jobs is installing new sfconfig rpm package, which is
not replacing the sfconfig.yaml file, so the new defaultl values for
logstash are not properly set.
This commit is getting logstash configuration from new
sfconfig.yaml.rpmnew file. If the file does not exist, it will
apply values that are set in sfconfig.yaml file.

[1] https://softwarefactory-project.io/r/#/c/20778/

Change-Id: Ie0681d35238b37d19d89b40acbbecb52eabdac14