Improve Kibana configuration
After login to the Kibana, it shows a popup to "select your tenant".
In the current version of Software Factory Project, we don't
support multitenancy, so we enable that option to hide the popup, but
still there is only one tenant to choose - "Global".
Allso allow kibana_viewer user to read the zuul.* index pattern

This commit provides a workaround for Kibana [1] that show a popup
to choose the default tenant where only one is enabled.

[1] https://github.com/opendistro-for-elasticsearch/security-kibana-plugin/issues/728

Change-Id: I64edb91a55e316c1801cb1856f9145e1e2070736