gerrit: prioritize custom commentlinks
Currently, when configuring gerrit, we add some default commentlinks
and then add any custom commentlinks provided by the operator. This
can cause an issue where the custom commentlinks are overriden by
the default ones, and it is happening on RDO with [1].

This commit moves the custom commentlinks above the default ones in
the gerrit.config template, so they always have priority.

[1] - https://github.com/rdo-infra/review.rdoproject.org-config/blob/master/gerrit/commentlinks.yaml#L4-L6

Change-Id: I399d913c8564e3cc770db38d9de5ca96c00ac247