Enable using patched code in containers
If a path is passed as an argument to /usr/local/bin/container-*.sh,
this path will replace the application library in the container.

This works for zuul and nodepool containers only; this is meant as
a facility for developers needing to test or apply a change rapidly.

Change-Id: Ib5fc324e2b800471eecb7bd07b9160ab9a8037cc
Example: /usr/local/bin/container-zuul-executor.sh /path/to/patchedzuul