nodepool: run the nodepool builder service as root
When using unprivileged users with uidmap, it seems like the cap_dac_override
does not work as expected, and when dib uses sudo, it can't access /etc/nodepool.
One solution would be to chown the group to 4000 on the host, and chmod 770
so that the root user in the container can read the files.
But since nodepool-builder is expected to run sudo, we might as well starts the
service as root directly and avoid tricky file modes.

Change-Id: I25f5a5650fc175e0b680d126509b98b34f0bc8a3