upgrade: recreate the container when needed
This change enables upgrading the containerized service
by recreating the container when necessary.

To do that we need to manage container creation at the install
phase so that the upgrade can recreate the container before the
setup phase. Therefor, this change moves the user creation from
the setup to the install.

Change-Id: Ie3d0db300d039474d2164e8314f4eb138d312cf9