Add a zuul-client wrapper, config generator utility
Add a wrapper to zuul-client's container on the zuul scheduler nodes.
The wrapper is preconfigured to use /etc/zuul/zuul.conf when
auto-generating tokens, and is pointing at {{ zuul_public_url }}. This
makes using zuul-client seamless with regard to the old zuul admin CLI,
except for the new executable's name.

Add a utility script to generate a zuul-client configuration
automatically.

Modify the zuul-restart playbook to use zuul-client for enqueues.

Depends-On: https://softwarefactory-project.io/r/c/containers/+/24809

Change-Id: I9d90d68291d86dc58e11eb9765b8235aee644c48