sf-container - when service file updated then restart service
This change attempts to fix an issue where when a service file is
installed then sf-container attempt to reload services than cannot
be reloaded (ExecReload is only added for the zuul-scheduler service).

When the service file is updated services must be restarted execpt
zuul/nodepool.

Also this change remove a double call to the contrainer creation script.
Indeed it might be called in the sf-container/upgrade.yml task but in
any cases it will be called in sf-container/setup.yml that will
happen after the sf-container/upgrade.yml.

And in sf-hound role we remove a call to legacy tasks/check_version.yml.

Change-Id: I2beff184c69ed5f801c37c5f3f6282784ecd055e