Improve ci-log-processing role; drop old elasticsearch user credentials
The logscraper and logsender tools are using same directory: first one
is downloading logs to it, second parse and send. It means both needs to
have rw access to that directory and the dir should not got permissions
for others.

Also drop elasticsearch_* credentials, which means that during the setup
all passwords for OpenSearch and OpenSearch Dashboards will be
regenerated.

Depends-on: https://softwarefactory-project.io/r/c/software-factory/sf-ci/+/27007

Change-Id: Ia9ad0e056d91f469cbd9d8cbbc7d316b1e12f1f8