Set log_size_max size for podman logs
The service logs after a while can be really huge.
This change is limiting log file size to 1GB.
The feature has been added into the podman containers.conf file in
podman 2.2.0 release [1], but on Centos 7, version is below 2.2.0.
According to the libpod.conf man [2], that option should be also
available in podman 1.6.4, but it is located in libpod.conf file.
More info [3].

[1] https://github.com/containers/podman/releases/tag/v2.2.0
[2] https://manpages.debian.org/unstable/podman/libpod.conf.5.en.html
[3] https://github.com/unifi-utilities/unifios-utilities/issues/100

Depends-On: https://softwarefactory-project.io/r/c/software-factory/sf-ci/+/28529

Change-Id: Ia6071e5214644bdd126cf696cd437c140fa95c94